Update image source paths and button styling in ReasonChoose and JapanElevator components
