--- src/app/[lang]/(support)/thang-may-gia-dinh/page.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function page() {
-  return <div>page</div>
-}
