+++ src/app/[lang]/(support)/thang-may-nhat-ban/[...slug]/page.jsx
@@ -1,0 +1,10 @@
+import FamilyElevator from '@/sections/thang-may-page-details/FamilyElevator'
+import React from 'react'
+
+const page = () => {
+  return (
+    <FamilyElevator/>
+  )
+}
+
+export default page
