+++ src/components/platFormMbJapanElevator/styles.css
@@ -1,0 +1,4 @@
+.slide_elevator_custom_mb .swiper-pagination-progressbar-fill{
+    background-color: #FE4127 !important;
+    border-radius: 99rem;
+}
