+++ src/components/platFormMobile/index.jsx
@@ -1,5 +1,8 @@
+'use client'
 import Content from '../platForm/content/Content'
 import Slider from './slider/Slider'
+import {Swiper, SwiperSlide} from 'swiper/react'
+import {Navigation, Pagination, Autoplay} from 'swiper/modules'
 function PlatFormMobile() {
   return (
     <section className='box-container-border bg-grey-600 pt-[1.69rem] pb-[3.19rem]'>
