+++ src/components/platFormMobile/slider/styles.css
@@ -1,4 +1,4 @@
-.slide_elevator_custom_mb .swiper-pagination-progressbar-fill{
+.slide_elevator_custom_mb_home .swiper-pagination-progressbar-fill{
     background-color: #C6A567 !important;
     border-radius: 99rem;
 }
