+++ src/components/sixReasons/index.jsx
@@ -27,9 +27,9 @@ const SixReasons = () => {
     },
   };
   return (
-    <section className='pl-[6.25rem] flex flex-col max-md:pl-0'>
+    <div className='pl-[6.25rem] flex flex-col max-md:pl-0'>
       <div>
-        <h2 className='font-SVNLagu text-[3.125rem] font-semibold leading-[140%] w-[49rem] mb-3 max-md:mb-2 max-md:px-4 max-md:w-full max-md:text-[1.5rem] max-md:order-1'>
+        <h2 className='font-SVNLagu text-[3.125rem] font-semibold leading-[140%] w-[49rem] mb-3 max-md:mb-2 max-md:px-4 max-md:text-[1.875rem] max-md:order-1 max-md:w-[20.9375rem]'>
           6 LÝ DO LỰA CHỌN THANG MÁY NHẬT BẢN SANYO YUSOKI
         </h2>
         <span className='w-[35rem] font-Iciel block text-base leading-[150%] font-normal mb-[0.5rem]  max-md:text-[0.875rem] max-md:text-justify max-md:px-4 max-md:w-full max-md:mb-6'>
@@ -86,7 +86,7 @@ const SixReasons = () => {
           ))}
         </Swiper>
       </div>
-    </section>
+    </div>
   )
 }
