+++ src/sections/thang-may-page-details/FamilyElevator.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import HeaderDetailElevator from './headerDetailElevator'
+
+const FamilyElevator = () => {
+  return (
+    <main>
+      <HeaderDetailElevator/>
+    </main>
+  )
+}
+
+export default FamilyElevator
